requirements.txt
aliyun-python-sdk-alinlp
tqdm
baidu-aip
chardet